# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProcessEmailBody {
  /** Email subject */
  subject: string;
  /** Email body text */
  body: string;
  /** Name of the sender */
  senderName?: string;
  /** Email address of the sender */
  senderEmail?: string;
}

export type EmailAnalysisResultSentiment =
  (typeof EmailAnalysisResultSentiment)[keyof typeof EmailAnalysisResultSentiment];

export const EmailAnalysisResultSentiment = {
  Positive: "Positive",
  Negative: "Negative",
  Neutral: "Neutral",
} as const;

export type EmailAnalysisResultIntent =
  (typeof EmailAnalysisResultIntent)[keyof typeof EmailAnalysisResultIntent];

export const EmailAnalysisResultIntent = {
  Support_Request: "Support Request",
  Feedback: "Feedback",
  Sales_Inquiry: "Sales Inquiry",
  General_Inquiry: "General Inquiry",
  Complaint: "Complaint",
  Other: "Other",
} as const;

export type EmailAnalysisResultArea =
  (typeof EmailAnalysisResultArea)[keyof typeof EmailAnalysisResultArea];

export const EmailAnalysisResultArea = {
  Product: "Product",
  Billing: "Billing",
  Technical: "Technical",
  Sales: "Sales",
  General: "General",
  Other: "Other",
} as const;

export type EmailAnalysisResultSource =
  | (typeof EmailAnalysisResultSource)[keyof typeof EmailAnalysisResultSource]
  | null;

export const EmailAnalysisResultSource = {
  manual: "manual",
  gmail: "gmail",
} as const;

export interface EmailAnalysisResult {
  id: number;
  subject: string;
  body: string;
  senderName?: string | null;
  senderEmail?: string | null;
  sentiment: EmailAnalysisResultSentiment;
  intent: EmailAnalysisResultIntent;
  area: EmailAnalysisResultArea;
  generatedReply: string;
  source?: EmailAnalysisResultSource;
  gmailMessageId?: string | null;
  replySent?: boolean | null;
  createdAt: string;
}

export interface EmailHistoryResponse {
  emails: EmailAnalysisResult[];
  total: number;
}

export type EmailStatsSentimentBreakdown = {
  Positive: number;
  Negative: number;
  Neutral: number;
};

export type EmailStatsIntentBreakdown = { [key: string]: number };

export type EmailStatsAreaBreakdown = { [key: string]: number };

export type EmailStatsRecentActivityItem = {
  date: string;
  count: number;
};

export interface EmailStats {
  total: number;
  sentimentBreakdown: EmailStatsSentimentBreakdown;
  intentBreakdown: EmailStatsIntentBreakdown;
  areaBreakdown: EmailStatsAreaBreakdown;
  recentActivity: EmailStatsRecentActivityItem[];
}

export interface AuthUrlResponse {
  url: string;
}

export interface GmailStatusResponse {
  connected: boolean;
  email?: string | null;
  name?: string | null;
}

export interface AutoRespondResult {
  success: boolean;
  message: string;
  analysis?: EmailAnalysisResult | null;
}

export interface SuccessResponse {
  success: boolean;
  message?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetEmailHistoryParams = {
  limit?: number;
  offset?: number;
};