# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Analyze email sentiment, intent, category and generate an AI reply
 * @summary Process and analyze an email with AI
 */
export const ProcessEmailBody = zod.object({
  subject: zod.string().describe("Email subject"),
  body: zod.string().describe("Email body text"),
  senderName: zod.string().optional().describe("Name of the sender"),
  senderEmail: zod.string().optional().describe("Email address of the sender"),
});

export const ProcessEmailResponse = zod.object({
  id: zod.number(),
  subject: zod.string(),
  body: zod.string(),
  senderName: zod.string().nullish(),
  senderEmail: zod.string().nullish(),
  sentiment: zod.enum(["Positive", "Negative", "Neutral"]),
  intent: zod.enum([
    "Support Request",
    "Feedback",
    "Sales Inquiry",
    "General Inquiry",
    "Complaint",
    "Other",
  ]),
  area: zod.enum([
    "Product",
    "Billing",
    "Technical",
    "Sales",
    "General",
    "Other",
  ]),
  generatedReply: zod.string(),
  source: zod.enum(["manual", "gmail"]).nullish(),
  gmailMessageId: zod.string().nullish(),
  replySent: zod.boolean().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * Returns list of previously processed emails with their analysis
 * @summary Get email processing history
 */
export const getEmailHistoryQueryLimitDefault = 20;
export const getEmailHistoryQueryOffsetDefault = 0;

export const GetEmailHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getEmailHistoryQueryLimitDefault),
  offset: zod.coerce.number().default(getEmailHistoryQueryOffsetDefault),
});

export const GetEmailHistoryResponse = zod.object({
  emails: zod.array(
    zod.object({
      id: zod.number(),
      subject: zod.string(),
      body: zod.string(),
      senderName: zod.string().nullish(),
      senderEmail: zod.string().nullish(),
      sentiment: zod.enum(["Positive", "Negative", "Neutral"]),
      intent: zod.enum([
        "Support Request",
        "Feedback",
        "Sales Inquiry",
        "General Inquiry",
        "Complaint",
        "Other",
      ]),
      area: zod.enum([
        "Product",
        "Billing",
        "Technical",
        "Sales",
        "General",
        "Other",
      ]),
      generatedReply: zod.string(),
      source: zod.enum(["manual", "gmail"]).nullish(),
      gmailMessageId: zod.string().nullish(),
      replySent: zod.boolean().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * Returns aggregated stats for sentiment, intent, and category breakdowns
 * @summary Get email analysis statistics
 */
export const GetEmailStatsResponse = zod.object({
  total: zod.number(),
  sentimentBreakdown: zod.object({
    Positive: zod.number(),
    Negative: zod.number(),
    Neutral: zod.number(),
  }),
  intentBreakdown: zod.record(zod.string(), zod.number()),
  areaBreakdown: zod.record(zod.string(), zod.number()),
  recentActivity: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * Returns the OAuth URL for Gmail authentication
 * @summary Get Gmail OAuth authorization URL
 */
export const GetGmailAuthUrlResponse = zod.object({
  url: zod.string(),
});

/**
 * Returns whether Gmail is connected and connected email
 * @summary Get Gmail connection status
 */
export const GetGmailStatusResponse = zod.object({
  connected: zod.boolean(),
  email: zod.string().nullish(),
  name: zod.string().nullish(),
});

/**
 * Reads latest unread Gmail, analyzes it, and sends AI-generated reply
 * @summary Auto-respond to latest unread Gmail
 */
export const GmailAutorespondResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  analysis: zod
    .object({
      id: zod.number(),
      subject: zod.string(),
      body: zod.string(),
      senderName: zod.string().nullish(),
      senderEmail: zod.string().nullish(),
      sentiment: zod.enum(["Positive", "Negative", "Neutral"]),
      intent: zod.enum([
        "Support Request",
        "Feedback",
        "Sales Inquiry",
        "General Inquiry",
        "Complaint",
        "Other",
      ]),
      area: zod.enum([
        "Product",
        "Billing",
        "Technical",
        "Sales",
        "General",
        "Other",
      ]),
      generatedReply: zod.string(),
      source: zod.enum(["manual", "gmail"]).nullish(),
      gmailMessageId: zod.string().nullish(),
      replySent: zod.boolean().nullish(),
      createdAt: zod.coerce.date(),
    })
    .nullish(),
});

/**
 * Disconnects the Gmail account and removes stored tokens
 * @summary Revoke Gmail access
 */
export const RevokeGmailAccessResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().nullish(),
});