# lib/api-zod/src/generated/types/emailAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAnalysisResultArea } from "./emailAnalysisResultArea";
import type { EmailAnalysisResultIntent } from "./emailAnalysisResultIntent";
import type { EmailAnalysisResultSentiment } from "./emailAnalysisResultSentiment";
import type { EmailAnalysisResultSource } from "./emailAnalysisResultSource";

export interface EmailAnalysisResult {
  id: number;
  subject: string;
  body: string;
  senderName?: string | null;
  senderEmail?: string | null;
  sentiment: EmailAnalysisResultSentiment;
  intent: EmailAnalysisResultIntent;
  area: EmailAnalysisResultArea;
  generatedReply: string;
  source?: EmailAnalysisResultSource;
  gmailMessageId?: string | null;
  replySent?: boolean | null;
  createdAt: Date;
}