# lib/api-zod/src/generated/types/emailHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAnalysisResult } from "./emailAnalysisResult";

export interface EmailHistoryResponse {
  emails: EmailAnalysisResult[];
  total: number;
}