# lib/api-zod/src/generated/types/gmailStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */

export interface GmailStatusResponse {
  connected: boolean;
  email?: string | null;
  name?: string | null;
}