# lib/api-zod/src/generated/types/processEmailBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Autoresponder API
 * OpenAPI spec version: 0.1.0
 */

export interface ProcessEmailBody {
  /** Email subject */
  subject: string;
  /** Email body text */
  body: string;
  /** Name of the sender */
  senderName?: string;
  /** Email address of the sender */
  senderEmail?: string;
}